mx-kz-ytml-2/kz-ytml-1/wz-hfmcp-1/WX/curs/cose1/pyproject.toml
[project]
name = "cose1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio[mcp]>=5.32.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "smolagents[toolkit]>=1.17.0",
    "tavily-python>=0.7.4",
]
